COSBETA-125: Review markups
