+++ cosmetics-web/app/models/notification.rb
@@ -21,7 +21,8 @@ class Notification < ApplicationRecord
   before_save :add_import_country, if: :will_save_change_to_import_country?
 
   validate :all_required_attributes_must_be_set
-  validates :cpnp_reference, uniqueness: { scope: :responsible_person, message: "Notification duplicated" }
+  validates :cpnp_reference, uniqueness: { scope: :responsible_person, message: "Notification duplicated" },
+            :allow_nil => true
 
   # rubocop:disable Metrics/BlockLength
   aasm whiny_transitions: false, column: :state do
